+++ client/src/routes/deleteAccount.module.scss
@@ -1,0 +1,57 @@
+@charset "utf-8";
+@import "../scss/variables.scss";
+
+.deleteAccount_wrapper {
+  text-align: center;
+  margin-left: 20%;
+  width: 1000px;
+  height: 500px;
+  border: $main200 1px solid;
+
+  .info_wrapper {
+    margin-left: 18.5%;
+    align-items: center;
+    text-align: center;
+    margin-top: 10%;
+    display: flex;
+    font-size: 14px;
+  }
+
+  .warning {
+    margin-top: 10px;
+    font-size: 14px;
+  }
+  .checkInput_wrapper {
+    margin-top: 5%;
+    .input_wrapper {
+      margin-left: 26%;
+      width: 500px;
+      height: 50px;
+      border: $main100 1px solid;
+    }
+  }
+  .button_wrapper {
+    margin-left: 26%;
+    .normal_btn {
+      width: 500px;
+      height: 50px;
+      background-color: $main300;
+      margin-top: 10px;
+    }
+    .red_btn {
+      width: 500px;
+      height: 50px;
+      background-color: rgb(234, 145, 145);
+      margin-top: 10px;
+    }
+  }
+  .last_info {
+    margin-top: 20px;
+  }
+  .text_red {
+    color: red;
+  }
+  .text_default {
+    color: $main700;
+  }
+}
